+++ webcomponents/article/src/article.js
@@ -1,10 +1,21 @@
-import { ContentElementClass } from '@trvt/core';
+import { TrivetElement } from '@trvt/core';
 import articleCSS from './article.css' assert { type: 'css' };
 
-export class TrvtArticle extends ContentElementClass {
+const contentTemplate = `
+	<article>
+		<slot name="heading"></slot>
+		<slot name="intro"></slot>
+		<slot name="content"></slot>
+		<slot name="footer"></slot>
+		<slot></slot>
+	</article>
+`;
+
+export class TrvtArticle extends TrivetElement {
 	constructor() {
 		super();
-		this.shadowRoot.adoptedStyleSheets = [articleCSS];
+		this.shadowStyleSheets = articleCSS;
+		this.template = contentTemplate;
 	}
 }
 customElements.define('trvt-article', TrvtArticle);