+++ webcomponents/card/src/card.js
@@ -1,10 +1,21 @@
-import { ContentElementClass } from '@trvt/core';
+import { TrivetElement } from '@trvt/core';
 import cardCSS from './card.css' assert { type: 'css' };
 
-export class TrvtCard extends ContentElementClass {
+const contentTemplate = `
+	<article>
+		<slot name="heading"></slot>
+		<slot name="intro"></slot>
+		<slot name="content"></slot>
+		<slot name="footer"></slot>
+		<slot></slot>
+	</article>
+`;
+
+export class TrvtCard extends TrivetElement {
 	constructor() {
 		super();
-		this.shadowRoot.adoptedStyleSheets = [cardCSS];
+		this.shadowStyleSheets = [cardCSS];
+		this.template = contentTemplate;
 	}
 }
 customElements.define('trvt-card', TrvtCard);