+++ webcomponents/core/src/defineElement.js
@@ -1,4 +1,0 @@
-const defineElement = (tag,compClass) => {
-	window.customElements.get(tag) || window.customElements.define(tag, compClass);
-}
-export { defineElement }