+++ webcomponents/contenttoggle/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @trvt/contenttoggle
 
+## 0.4.1
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @trvt/core@0.9.0
+
 ## 0.4.0
 
 ### Minor Changes