+++ webcomponents/datatable/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @trvt/datatable
 
+## 0.3.3
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @trvt/core@0.9.0
+
 ## 0.3.2
 
 ### Patch Changes