+++ webcomponents/dragscroller/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @trvt/dragscroller
 
+## 0.3.2
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @trvt/core@0.9.0
+
 ## 0.3.1
 
 ### Patch Changes