+++ webcomponents/lists/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @trvt/lists
 
+## 0.4.2
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @trvt/core@0.9.0
+
 ## 0.4.1
 
 ### Patch Changes